implement crypto_sign_keypair wrapper
crypto_sign code is moved from lib to it's own module, and a
crypto_sign_keypair prototype is implemented based off the tweetnacl
c++ bindings.
